+++ package.json
@@ -107,18 +107,18 @@
     "openid-client": "^5.7.1",
     "passport": "^0.7.0",
     "passport-oauth2": "^1.8.0",
-    "pino": "^9.8.0",
-    "redis": "^5.8.0",
+    "pino": "^9.9.0",
+    "redis": "^5.8.1",
     "sass": "^1.90.0",
     "superagent": "^10.2.3",
     "web-did-resolver": "^2.0.30",
-    "zod": "^4.0.15"
+    "zod": "^4.0.17"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.8.1",
     "@commitlint/config-conventional": "^19.8.1",
     "@jgoz/esbuild-plugin-typecheck": "^4.0.3",
-    "@ministryofjustice/eslint-config-hmpps": "^0.0.2",
+    "@ministryofjustice/eslint-config-hmpps": "^0.0.3",
     "@tsconfig/node22": "^22.0.2",
     "@types/bunyan": "^1.8.11",
     "@types/bunyan-format": "^0.2.9",
@@ -141,12 +141,12 @@
     "aws-sdk-client-mock": "^4.1.0",
     "chokidar": "^3.6.0",
     "concurrently": "^9.2.0",
-    "cypress": "^14.5.3",
+    "cypress": "^14.5.4",
     "cypress-multi-reporters": "^2.0.5",
     "dotenv": "^17.2.1",
     "dotenv-cli": "^10.0.0",
     "dotenv-expand": "^12.0.2",
-    "esbuild": "^0.25.8",
+    "esbuild": "^0.25.9",
     "esbuild-plugin-clean": "^1.0.1",
     "esbuild-plugin-copy": "^2.1.1",
     "esbuild-plugin-manifest": "^1.0.5",
@@ -156,17 +156,17 @@
     "jest-html-reporter": "^4.3.0",
     "jest-junit": "^16.0.0",
     "jsonwebtoken": "^9.0.2",
-    "lint-staged": "^16.1.4",
+    "lint-staged": "^16.1.5",
     "mocha-junit-reporter": "^2.2.1",
-    "nock": "^14.0.9",
+    "nock": "^14.0.10",
     "npm-check-updates": "^18.0.2",
-    "pkgroll": "^2.15.2",
+    "pkgroll": "^2.15.3",
     "prettier": "^3.6.2",
     "prettier-plugin-jinja-template": "^2.1.0",
     "supertest": "^7.1.4",
     "ts-jest": "^29.4.1",
     "ts-node": "^10.9.2",
-    "tsx": "^4.20.3",
+    "tsx": "^4.20.4",
     "typescript": "^5.9.2"
   },
   "overrides": {
