+++ server/controllers/case-dashboard-controller.ts
@@ -7,15 +7,16 @@ const caseDashboardController = async (req: Request, res: Response, next: NextFu
     await initialiseBasicAuthentication(req, res, next)
 
     res.locals.pageTitle = 'Your case dashboard'
-    res.locals.backLink = '/case/search'
-    res.locals.caseReference = req.session.selectedUrn || req.session.selectedCrn
-    res.locals.pageTitleCaseReference = ` - Reference number: ${res.locals.caseReference}`
+    res.locals.backLink = '/'
+    // res.locals.caseReference = req.session.selectedUrn || req.session.selectedCrn
+    // res.locals.pageTitleCaseReference = ` - Reference number: ${res.locals.caseReference}`
 
-    if (res.locals.caseReference) {
-      res.render('pages/case/dashboard.njk')
-    } else {
-      res.redirect(paths.CASES.SEARCH)
-    }
+    // if (res.locals.caseReference) {
+    //   res.render('pages/case/dashboard.njk')
+    // } else {
+    //   res.redirect(paths.CASES.SEARCH)
+    // }
+    res.render('pages/case/dashboard.njk')
   } catch (error) {
     next(error)
   }
