+++ server/views/pages/case/confirm-case.njk
@@ -35,7 +35,7 @@
   <div class="govuk-button-group">
     {{ govukButton({
       text: "Continue",
-      href: "/case/dashboard"
+      href: "/case/court-information"
     }) }}
 
     <a class="govuk-link" href="/case/search">Cancel</a>
