+++ server/views/pages/case/dashboard.njk
@@ -21,7 +21,7 @@
 
   }) }}
 
-  <p class="govuk-body">Crime reference number: <strong>{{ caseReference }}</strong></p>
+{#  <p class="govuk-body">Crime reference number: <strong>{{ caseReference }}</strong></p>#}
 
   <div class="govuk-grid-row govuk-!-margin-bottom-6">
     <div class="govuk-grid-column-one-third">
@@ -32,7 +32,7 @@
         {{ actionCard(
           "View court information",
           "See information about the court case, including date, location, your requirements and verdicts.",
-          "/case/court-information-2",
+          "/case/search",
           "View court information",
           "default",
           0
