+++ server/views/pages/index.njk
@@ -36,8 +36,8 @@
 
       {% if useOneLogin %}
         {{ govukButton({
-          text: "Make a request for authentication",
-          href: "/signed-in",
+          text: "Start now",
+          href: "/case/dashboard",
           isStartButton: true
         }) }}
       {% else %}
