+++ server/views/partials/header.njk
@@ -13,9 +13,9 @@
       text: 'Home',
       id: 'home'
     },{
-      href: '/case/search',
-      text: 'Case search',
-      id: 'case-search'
+      href: '/case/dashboard',
+      text: 'Dashboard',
+      id: 'case-dashboard'
     }]
   }) }}
 {% else %}
